refactor code for better read
